linking system part 1

+++ src/main/java/com/ustudents/engine/core/json/JsonReader.java
@@ -260,11 +260,12 @@ public class JsonReader {
             contentBuilder.append(currentCharacter);
         }
         String number = contentBuilder.toString();
+
         if (number.contains(".")) {
             return Double.parseDouble(number);
         }
 
-        return Integer.parseInt(number);
+        return Long.parseLong(number);
     }
 
     /**
